Update instructions to make less confusing
I found it difficult following the instructions, either because some
steps were missing, or there is too much detail when not required.

A quick start guide should not document "what we would like to see
in the future." Rather it should describe how to do it right now
with existing tools.

This commit fixes some of the problems that I had.